moved ListLevel to a separate file
