+++ src/course/drawer/Drawer.tsx
@@ -1,12 +1,11 @@
-import React, {useEffect, useState} from "react";
-import {useHistory, useParams} from "react-router-dom";
+import React, {useContext, useEffect, useState} from "react";
+import {useHistory} from "react-router-dom";
 
 import { styled, useTheme, Theme, CSSObject } from '@mui/material/styles';
 import Box from '@mui/material/Box';
 import MuiDrawer from '@mui/material/Drawer';
 import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
 import CssBaseline from '@mui/material/CssBaseline';
 import IconButton from '@mui/material/IconButton';
 import MenuIcon from '@mui/icons-material/Menu';
@@ -15,16 +14,16 @@ import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import ListItem from '@mui/material/ListItem';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
-import {Home, ArrowDropUp, ArrowDropDown, Equalizer} from "@mui/icons-material";
-import ArrowRightIcon from '@mui/icons-material/ArrowRight';
+import {Home} from "@mui/icons-material";
 import QueryStatsIcon from '@mui/icons-material/QueryStats';
 
-import {Exercise} from "../models/courses";
-import {AppBarProfile} from "../header/Auth";
-import {Progress} from "../models/users";
-import {useStatusToStyledBackground} from "./colors";
-import AppBarNotifications from "../header/Notifications";
-import {Typography} from "@mui/material";
+import {Course, Exercise} from "../../models/courses";
+import {AppBarProfile} from "../../header/Auth";
+import {Progress} from "../../models/users";
+import AppBarNotifications from "../../header/Notifications";
+import LevelList from "./LevelList";
+import {onUserProgressUpdated} from "../../services/users";
+import {AuthContext} from "../../App";
 
 
 const drawerWidth = 240;
@@ -102,84 +101,21 @@ const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' 
     }),
 );
 
-function LevelList({levelNumber, exercises, progress, onItemSelected, isDrawerOpen, isSingleLevel}:
-                   {
-                       levelNumber: number,
-                       exercises: Exercise[],
-                       progress: { [key: string]: Progress },
-                       onItemSelected: (exerciseId: string) => void,
-                       isDrawerOpen: boolean,
-                       isSingleLevel: boolean
-                   }) {
-    const {exerciseId} = useParams<{ exerciseId?: string }>();
-    const isExerciseInLevel = exercises.filter(e => e.id === exerciseId).length > 0;
-    const [open, setOpen] = useState(isExerciseInLevel || isSingleLevel);
-    const statusToStyle = useStatusToStyledBackground();
 
-    useEffect(() => {
-        if( isSingleLevel )
-            setOpen(true);
-        else if( !isSingleLevel && open && !isExerciseInLevel ) {
-            setOpen(false);
-        }
-        else if( !open ) {
-            setOpen(isExerciseInLevel);
-        }
-        // open the level if the current exercise is in this level
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [exercises, isExerciseInLevel, isSingleLevel]);
-
-    const onLevelClicked = () => setOpen(!open);
-    const getStatusStyle = (id: string) => {
-        const status = id in progress ? progress[id].status : undefined;
-        if( !status )
-            return statusToStyle.undefined;
-        return statusToStyle[status];
-    }
-
-    let levelClass = statusToStyle.Solved;
-    for( const e of exercises) {
-        if( getStatusStyle(e.id) !== levelClass ) {
-            levelClass = statusToStyle.undefined;
-            break;
-        }
-    }
-
-    return <>
-        <List disablePadding>
-            {!isSingleLevel &&
-                <ListItem button key={`level-${levelNumber}`} onClick={onLevelClicked} className={levelClass}>
-                    <ListItemIcon>
-                        <Equalizer/>
-                        {!isDrawerOpen && <Typography variant="subtitle1">{levelNumber}</Typography>}
-                        {isDrawerOpen && <Typography variant="subtitle1">Level {levelNumber}</Typography>}
-                        {open ? <ArrowDropUp/> : <ArrowDropDown/>}
-                    </ListItemIcon>
-                </ListItem>
-            }
-            {open && exercises.map((ex, index) =>
-                <ListItem button key={ex.id} onClick={() => onItemSelected(ex.id)} className={getStatusStyle(ex.id)}>
-                    <ListItemIcon>{exerciseId === ex.id ? <ArrowRightIcon /> : <ListItemText primary={index + 1}/>}</ListItemIcon>
-                    <ListItemText primary={ex.title}/>
-                </ListItem>
-            )}
-        </List>
-    </>
-}
-
-
-function CourseDrawer({exercises, progress, onItemSelected, showRanking, onRankingClicked}:
+function CourseDrawer({course, exercises, onItemSelected, showRanking, onRankingClicked}:
                       {
+                          course: Course | null,
                           exercises: Exercise[],
-                          progress: { [key: string]: Progress},
                           onItemSelected: (exerciseId: string) => void,
                           showRanking: boolean,
                           onRankingClicked: () => void,
                       }) {
+    const auth = useContext(AuthContext);
     const theme = useTheme();
     const history = useHistory();
     const [open, setOpen] = useState(false);
     const [levels, setLevels] = useState<Exercise[][]>([]);
+    const [progress, setProgress] = useState<{[key: string]: Progress}>({});
 
     const handleDrawerOpen = () => setOpen(true);
     const handleDrawerClose = () => setOpen(false);
@@ -204,6 +140,14 @@ function CourseDrawer({exercises, progress, onItemSelected, showRanking, onRanki
 
         setLevels(levels);
     }, [exercises]);
+
+    useEffect(() => {
+        if( !auth.currentUserId || !course?.id )
+            return;
+
+        const unsubscribe = onUserProgressUpdated(auth.currentUserId, course.id, setProgress);
+        return () => unsubscribe();
+    }, [course?.id, auth]);
 
 
     return (
